+++ semana7/sistema-cadastro/sistema-cadastro-usuarios/public/favicon.ico
