+++ semana7/sistema-cadastro/sistema-cadastro-usuarios/public/manifest.json
