+++ semana7/sistema-cadastro/sistema-cadastro-usuarios/src/constants/constantesAxios.js
@@ -1,0 +1,7 @@
+export const baseUrl = "https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users"
+
+export const headers = {
+    headers: {
+    Authorization: "magali-silva-jackson"
+    }
+}
